Split off test utilities in own module
