+++ consumers/test/Test.hs
@@ -1,57 +1,25 @@
 module Main where
 
-import Control.Applicative ((<|>))
+import Control.Concurrent (threadDelay)
 import Control.Concurrent.STM
 import Control.Monad
-import Control.Monad.Base
 import Control.Monad.Catch
 import Control.Monad.IO.Class
-import Control.Monad.State.Strict
+import Control.Monad.RWS
 import Control.Monad.Time
-import Control.Monad.Trans.Control
+import Data.IORef
 import Data.Int
 import Data.Text qualified as T
 import Data.Time
 import Database.PostgreSQL.Consumers
 import Database.PostgreSQL.PQTypes
-import Database.PostgreSQL.PQTypes.Checks
 import Database.PostgreSQL.PQTypes.Model
 import Log
 import Log.Backend.StandardOutput
-import System.Environment
-import System.Exit
 import Test.HUnit qualified as T
 import Test.Tasty
 import Test.Tasty.HUnit
-
-data TestEnvSt = TestEnvSt
-  { teCurrentTime :: UTCTime
-  , teMonotonicTime :: Double
-  }
-
-type InnerTestEnv = StateT TestEnvSt (DBT (LogT IO))
-
-newtype TestEnv a = TestEnv {unTestEnv :: InnerTestEnv a}
-  deriving (Applicative, Functor, Monad, MonadLog, MonadDB, MonadThrow, MonadCatch, MonadMask, MonadIO, MonadBase IO, MonadState TestEnvSt)
-
-instance MonadBaseControl IO TestEnv where
-  type StM TestEnv a = StM InnerTestEnv a
-  liftBaseWith f = TestEnv $ liftBaseWith (\run -> f $ run . unTestEnv)
-  restoreM = TestEnv . restoreM
-
-instance MonadTime TestEnv where
-  currentTime = gets teCurrentTime
-  monotonicTime = gets teMonotonicTime
-
-modifyTestTime :: MonadState TestEnvSt m => (UTCTime -> UTCTime) -> m ()
-modifyTestTime modtime = modify (\te -> te {teCurrentTime = modtime . teCurrentTime $ te})
-
-runTestEnv :: ConnectionSourceM (LogT IO) -> Logger -> TestEnv a -> IO a
-runTestEnv connSource logger =
-  runLogT "consumers-test" logger defaultLogLevel
-    . runDBT connSource defaultTransactionSettings
-    . (\m' -> fst <$> runStateT m' (TestEnvSt (UTCTime (ModifiedJulianDay 0) 0) 0))
-    . unTestEnv
+import Util
 
 main :: IO ()
 main = do
@@ -63,136 +31,88 @@ allTests :: ConnectionSettings -> TestTree
 allTests connectionSource =
   testGroup
     "consumers"
-    [ testCase "can consume queue" (testPipeline connectionSource)
+    [ testCase "can grow the number of jobs ran concurrently" (testJobScheduleGrowth connectionSource)
     ]
 
 --------------------
 
-getConnectionString :: IO T.Text
-getConnectionString = do
-  connectionParamsString <- (<|>) <$> paramsFromGithub <*> paramsFromEnvironmentVariables
-  allArgs <- getArgs
-  case connectionParamsString of
-    Just params -> pure (stringFromParams params)
-    _ -> case allArgs of
-      connString : _args -> pure (T.pack connString)
-      [] -> printUsage *> exitFailure
-  where
-    printUsage = do
-      prog <- getProgName
-      putStrLn $ "Usage: " <> prog <> " <connection info string>"
-
-    paramsFromGithub =
-      lookupEnv "GITHUB_ACTIONS" >>= \case
-        Just "true" -> pure $ Just ("postgres", "postgres", "postgres")
-        _ -> pure $ Nothing
-    paramsFromEnvironmentVariables = do
-      variables <-
-        sequence
-          [ lookupEnv "PGHOST"
-          , lookupEnv "PGUSER"
-          , lookupEnv "PGDATABASE"
-          ]
-      case variables of
-        [Just host, Just user, Just database] -> pure $ Just (host, user, database)
-        _ -> pure $ Nothing
-    stringFromParams (host, user, database) =
-      (T.pack ("host=" <> host <> " user=" <> user <> " dbname=" <> database))
-
-testPipeline :: ConnectionSettings -> IO ()
-testPipeline connectionSettings = do
+testJobScheduleGrowth :: ConnectionSettings -> IO ()
+testJobScheduleGrowth connectionSettings = do
   let ConnectionSource connSource = simpleSource connectionSettings
-  withStdOutLogger $ \logger ->
-    runTestEnv connSource logger $ do
-      bracket createTables (const dropTables) $ \_ -> do
-        idleSignal <- liftIO newEmptyTMVarIO
-        putJob 10 >> commit
+  withStdOutLogger $ \logger -> do
+    let additionalColumns =
+          [ tblColumn
+              { colName = "countdown"
+              , colType = IntegerT
+              , colNullable = False
+              }
+          ]
+    runTestEnv connSource logger (TestSetup "test_job_schedule_growth" additionalColumns) $ do
+      consumerConfig <- getConsumerConfig
+      TestEnvSt {..} <- get
+      idleSignal <- liftIO newEmptyTMVarIO
+      putJob 10 >> commit
 
-        forM_ [1 .. 10 :: Int] $ \_ -> do
-          -- Move time forward 2hours, because jobs are scheduled 1 hour into future
-          modifyTestTime $ addUTCTime (2 * 60 * 60)
-          finalize
-            ( localDomain "process" $
-                runConsumerWithIdleSignal consumerConfig connSource idleSignal
-            )
-            $ do
-              waitUntilTrue idleSignal
-          currentTime >>= (logInfo_ . T.pack . ("current time: " ++) . show)
-
-        -- Each job creates 2 new jobs, so there should be 1024 jobs in table.
-        runSQL_ "SELECT COUNT(*) from consumers_test_jobs"
-        rowcount0 :: Int64 <- fetchOne runIdentity
-        -- Move time 2 hours forward
+      rowCountGrowth :: [Int64] <- forM [1 .. 10 :: Int] $ \_ -> do
+        -- Move time forward 2hours, because jobs are scheduled 1 hour into future
         modifyTestTime $ addUTCTime (2 * 60 * 60)
         finalize
           ( localDomain "process" $
               runConsumerWithIdleSignal consumerConfig connSource idleSignal
           )
-          $ do
-            waitUntilTrue idleSignal
-        -- Jobs are designed to double only 10 times, so there should be no jobs left now.
-        runSQL_ "SELECT COUNT(*) from consumers_test_jobs"
-        rowcount1 :: Int64 <- fetchOne runIdentity
-        liftIO $ T.assertEqual "Number of jobs in table after 10 steps is 1024" 1024 rowcount0
-        liftIO $ T.assertEqual "Number of jobs in table after 11 steps is 0" 0 rowcount1
+          $ waitUntilTrue idleSignal
+        currentTime >>= (logInfo_ . T.pack . ("current time: " ++) . show)
+
+        -- Each job creates 2 new jobs, so there should be 1024 jobs in table.
+        runSQL_ ("SELECT COUNT(*) from " <> raw teJobTableName)
+        fetchOne runIdentity
+
+      -- Move time 2 hours forward
+      modifyTestTime $ addUTCTime (2 * 60 * 60)
+      finalize
+        ( localDomain "process" $
+            runConsumerWithIdleSignal consumerConfig connSource idleSignal
+        )
+        $ waitUntilTrue idleSignal
+
+      -- Jobs are designed to double only 10 times, so there should be no jobs left now.
+      runSQL_ ("SELECT COUNT(*) from " <> raw teJobTableName)
+      rowcount1 :: Int64 <- fetchOne runIdentity
+      liftIO $ T.assertEqual "Number of jobs in table after 10 steps grows exponentially" [2, 4, 8, 16, 32, 64, 128, 256, 512, 1024] rowCountGrowth
+      liftIO $ T.assertEqual "Number of jobs in table after 11 steps is 0" 0 rowcount1
   where
-    waitUntilTrue tmvar = liftIO . atomically $ do
-      takeTMVar tmvar >>= \case
-        True -> pure ()
-        False -> retry
-
-    definitions = emptyDbDefinitions {dbTables = [consumersTable, jobsTable]}
-    -- NB: order of migrations is important.
-    migrations =
-      [ createTableMigration consumersTable
-      , createTableMigration jobsTable
-      ]
-
-    createTables :: TestEnv ()
-    createTables = do
-      migrateDatabase
-        defaultExtrasOptions
-        definitions
-        migrations
-      checkDatabase
-        defaultExtrasOptions
-        definitions
-
-    dropTables :: TestEnv ()
-    dropTables = do
-      migrateDatabase
-        defaultExtrasOptions
-        emptyDbDefinitions
-        [ dropTableMigration jobsTable
-        , dropTableMigration consumersTable
-        ]
-
-    consumerConfig =
-      ConsumerConfig
-        { ccJobsTable = "consumers_test_jobs"
-        , ccConsumersTable = "consumers_test_consumers"
-        , ccJobSelectors = ["id", "countdown"]
-        , ccJobFetcher = id
-        , ccJobIndex = \(i :: Int64, _ :: Int32) -> i
-        , ccNotificationChannel = Just "consumers_test_chan"
-        , -- select some small timeout
-          ccNotificationTimeout = 100 * 1000 -- 100 msec
-        , ccMaxRunningJobs = 20
-        , ccProcessJob = processJob
-        , ccOnException = handleException
-        , ccJobLogData = \(i, _) -> ["job_id" .= i]
-        }
+    getConsumerConfig :: TestEnv (ConsumerConfig TestEnv Int64 (Int64, Int32))
+    getConsumerConfig = do
+      TestEnvSt {..} <- get
+      pure $
+        ConsumerConfig
+          { ccJobsTable = teJobTableName
+          , ccConsumersTable = teConsumerTableName
+          , ccJobSelectors = ["id", "countdown"]
+          , ccJobFetcher = id
+          , ccJobIndex = \(i :: Int64, _ :: Int32) -> i
+          , ccNotificationChannel = Just teNotificationChannel
+          , -- select some small timeout
+            ccNotificationTimeout = 100 * 1000 -- 100 msec
+          , ccMaxRunningJobs = 20
+          , ccProcessJob = processJob
+          , ccOnException = handleException
+          , ccJobLogData = \(i, _) -> ["job_id" .= i]
+          , ccMutexColumn = Nothing
+          }
 
     putJob :: Int32 -> TestEnv ()
     putJob countdown = localDomain "put" $ do
+      TestEnvSt {..} <- get
       now <- currentTime
       runSQL_ $
-        "INSERT INTO consumers_test_jobs "
+        "INSERT INTO "
+          <> raw teJobTableName
           <> "(run_at, finished_at, reserved_by, attempts, countdown) "
           <> "VALUES (" <?> now
           <> " + interval '1 hour', NULL, NULL, 0, " <?> countdown
           <> ")"
-      notify "consumers_test_chan" ""
+      notify teNotificationChannel ""
 
     processJob :: (Int64, Int32) -> TestEnv Result
     processJob (_idx, countdown) = do
@@ -202,93 +122,11 @@ testPipeline connectionSettings = do
         commit
       pure (Ok Remove)
 
-    handleException :: SomeException -> (Int64, Int32) -> TestEnv Action
-    handleException _ _ = pure . RerunAfter $ imicroseconds 500000
+waitUntilTrue :: MonadIO m => TMVar Bool -> m ()
+waitUntilTrue tmvar = liftIO . atomically $ do
+  takeTMVar tmvar >>= \case
+    True -> pure ()
+    False -> retry
 
-jobsTable :: Table
-jobsTable =
-  tblTable
-    { tblName = "consumers_test_jobs"
-    , tblVersion = 1
-    , tblColumns =
-        [ tblColumn
-            { colName = "id"
-            , colType = BigSerialT
-            , colNullable = False
-            }
-        , tblColumn
-            { colName = "run_at"
-            , colType = TimestampWithZoneT
-            , colNullable = True
-            }
-        , tblColumn
-            { colName = "finished_at"
-            , colType = TimestampWithZoneT
-            , colNullable = True
-            }
-        , tblColumn
-            { colName = "reserved_by"
-            , colType = BigIntT
-            , colNullable = True
-            }
-        , tblColumn
-            { colName = "attempts"
-            , colType = IntegerT
-            , colNullable = False
-            }
-        , -- The only non-obligatory field:
-          tblColumn
-            { colName = "countdown"
-            , colType = IntegerT
-            , colNullable = False
-            }
-        ]
-    , tblPrimaryKey = pkOnColumn "id"
-    , tblForeignKeys =
-        [ (fkOnColumn "reserved_by" "consumers_test_consumers" "id")
-            { fkOnDelete = ForeignKeySetNull
-            }
-        ]
-    }
-
-consumersTable :: Table
-consumersTable =
-  tblTable
-    { tblName = "consumers_test_consumers"
-    , tblVersion = 1
-    , tblColumns =
-        [ tblColumn
-            { colName = "id"
-            , colType = BigSerialT
-            , colNullable = False
-            }
-        , tblColumn
-            { colName = "name"
-            , colType = TextT
-            , colNullable = False
-            }
-        , tblColumn
-            { colName = "last_activity"
-            , colType = TimestampWithZoneT
-            , colNullable = False
-            }
-        ]
-    , tblPrimaryKey = pkOnColumn "id"
-    }
-
-createTableMigration :: MonadDB m => Table -> Migration m
-createTableMigration tbl =
-  Migration
-    { mgrTableName = tblName tbl
-    , mgrFrom = 0
-    , mgrAction = StandardMigration $ do
-        createTable True tbl
-    }
-
-dropTableMigration :: Table -> Migration m
-dropTableMigration tbl =
-  Migration
-    { mgrTableName = tblName tbl
-    , mgrFrom = 1
-    , mgrAction = DropTableMigration DropTableRestrict
-    }
+handleException :: SomeException -> k -> TestEnv Action
+handleException _ _ = pure . RerunAfter $ imicroseconds 500000
